pyproject.toml
[project]
name = "extraction-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "boto3>=1.40.48",
    "fastapi[standard]>=0.118.0",
    "html2text>=2025.4.15",
    "markitdown[all]>=0.1.3",
    "openai>=2.2.0",
    "pdfreader>=0.1.15",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "unstructured[all-docs,pdf]>=0.18.15",
    "uvicorn>=0.37.0",
]
